+++ lib/utils.js
@@ -3,59 +3,9 @@ module.exports = (() => {
   const os = require('os');
   const fs = require('fs');
   const path = require('path');
-  const execSync = require('child_process').execSync;
-  const spawn = require('child_process').spawn;
-  const promptly = require('promptly');
 
-  const MASTER_BRANCH = {
-    version: 'master',
-    feature: '',
-    ticketNumber: '',
-    description: '',
-    isBuildable: false
-  };
-
-  const BRANCH_TYPES = {
-    VERSION: {
-      name: 'version',
-      matches: (parsedBranch) => {
-        return _.isEmpty(parsedBranch.feature) && _.isEmpty(parsedBranch.description);
-      },
-      getBaseBranch: () => {
-        return MASTER_BRANCH;
-      }
-    },
-    FEATURE: {
-      name: 'feature',
-      matches: (parsedBranch) => {
-        return !_.isEmpty(parsedBranch.feature) && _.isEmpty(parsedBranch.description);
-      },
-      getBaseBranch: (parsedBranch) => {
-        return {
-          version: parsedBranch.version,
-          feature: '',
-          ticketNumber: '',
-          description: '',
-          isBuildable: false
-        }
-      }
-    },
-    DEV: {
-      name: 'dev',
-      matches: (parsedBranch) => {
-        return !_.isEmpty(parsedBranch.description);
-      },
-      getBaseBranch: (parsedBranch) => {
-        return {
-          version: parsedBranch.version,
-          feature: parsedBranch.feature,
-          ticketNumber: '',
-          description: '',
-          isBuildable: false
-        }
-      }
-    }
-  };
+  const execution = require('./utils/execution');
+  const branches = require('./utils/branches');
 
   const SPLITTER = '€$£';
 
@@ -129,45 +79,6 @@ module.exports = (() => {
     }
   };
 
-  const execute = command => {
-    return execSync(command).toString();
-  };
-
-  const executeAndPipe = (command, arguments) => {
-    spawn(command, arguments, { stdio: 'inherit' });
-  };
-
-  const print = console.log;
-
-  const exit = (message, exitCode) => {
-    print(exitCode === 0 ? message.green : message.red);
-    process.exit(exitCode);
-  };
-
-  const promisifiedPrompt = (message, options) => {
-    return new Promise((resolve, reject) => {
-      promptly.prompt(message, options, (error, value) => {
-        return error
-          ? reject(error)
-          : resolve(value);
-      });
-    });
-  };
-
-  const yesNoPrompt = (message, callback) => {
-    const options = {
-      default: 'n',
-      validator: choice => choice === 'y'
-    };
-    promptly.prompt(`${message} (y/n)`, options, (error, value) => {
-      if (error) {
-        throw error;
-      }
-
-      callback(value);
-    });
-  };
-
   const mergeArrayCustomizer = (seed, otherSource) => {
     if (_.isArray(seed)) {
       return seed.concat(otherSource);
@@ -175,7 +86,7 @@ module.exports = (() => {
   };
 
   const getTopLevel = () => {
-    const unsanitizedTopLevel = execSync('git rev-parse --show-toplevel');
+    const unsanitizedTopLevel = execution.execute('git rev-parse --show-toplevel');
     return _.replace(unsanitizedTopLevel, /\n/, '');
   };
 
@@ -185,7 +96,7 @@ module.exports = (() => {
 
   const isDirty = () => {
     try {
-      execSync('git diff --no-ext-diff --quiet --exit-code');
+      execution.execute('git diff --no-ext-diff --quiet --exit-code');
       return false;
     } catch (error) {
       return true;
@@ -194,7 +105,7 @@ module.exports = (() => {
 
   const hasStagedChanges = () => {
     try {
-      execSync('git diff-index --cached --quiet HEAD --');
+      execution.execute('git diff-index --cached --quiet HEAD --');
       return false;
     } catch (error) {
       return true;
@@ -203,103 +114,11 @@ module.exports = (() => {
 
   const hasUnstagedChanges = () => {
     try {
-      execSync('[ -n "$(git ls-files --others --exclude-standard)" ]');
+      execution.execute('[ -n "$(git ls-files --others --exclude-standard)" ]');
       return true;
     } catch (error) {
       return false;
     }
-  };
-
-  const searchForLocalBranch = regex => {
-    const allBranches = execute(`git branch 2> /dev/null`);
-    return _(allBranches.split('\n'))
-      .map(branch => _.replace(branch, /^[* ] /, ''))
-      .filter(branch => regex.test(branch))
-      .value();
-  };
-
-  const getCurrentBranch = () => {
-    const allBranches = execute(`git branch 2> /dev/null`);
-    return _(allBranches.split('\n'))
-      .filter(branch => branch.startsWith('*'))
-      .map(branch => _.replace(branch, '* ', ''))
-      .first();
-  };
-
-  const parseBranchName = (branch) => {
-    const branchName = branch || getCurrentBranch();
-    const branchRegex = /^(([0-9]+\.[0-9]+\.[0-9]+(\.[0-9]+)?)|master)(_([^#]+)?#([^_]+))?(_([0-9]+))?(_([^_]+))?$/;
-    if (branchName === 'master') {
-      return {
-        version: 'master',
-        feature: '',
-        ticketNumber: '',
-        description: '',
-        isBuildable: false
-      };
-
-    } else if (!branchRegex.test(branchName)) {
-      throw Error(`The branch name ${branchName} is not valid. Please read the specs at https://github.com/quilicicf/Gut/blob/master/specs/specs.md`.red);
-
-    }
-
-    const branchFragments = branchRegex.exec(branchName);
-    const parsedBranch = {
-      version: branchFragments[ 1 ],
-      feature: branchFragments[ 6 ] || '',
-      ticketNumber: branchFragments[ 8 ] || '',
-      description: branchFragments[ 10 ] || '',
-      isBuildable: branchFragments[ 5 ] === 'build'
-    };
-
-    const parsedBranchType = _.find(BRANCH_TYPES, (branchType) => {
-      return branchType.matches(parsedBranch);
-    });
-
-    parsedBranch.type = parsedBranchType.name;
-    parsedBranch.baseBranch = buildBranchName(parsedBranchType.getBaseBranch(parsedBranch));
-
-    return parsedBranch;
-  };
-
-  const buildBranchName = parsedBranch => {
-    const versionFragment = parsedBranch.version;
-    const featureDescription = parsedBranch.feature || '';
-    const featureFragment = featureDescription
-      ? `${parsedBranch.isBuildable ? BUILDABLE_BRANCH_TAG : ''}#${featureDescription}`
-      : '';
-    const ticketNumberFragment = '' + (parsedBranch.ticketNumber || '');
-    const descriptionFragment = parsedBranch.description || '';
-
-    return _([ versionFragment, featureFragment, ticketNumberFragment, descriptionFragment ])
-      .reject(fragment => _.isEmpty(fragment))
-      .join('_');
-  };
-
-  const isMasterBranch = parsedBranch => {
-    return parsedBranch.version === 'master'
-      && !parsedBranch.feature
-      && !parsedBranch.ticketNumber
-      && !parsedBranch.description;
-  };
-
-  const isVersionBranch = parsedBranch => {
-    return parsedBranch.version // TODO: semver-check dat
-      && !parsedBranch.feature
-      && !parsedBranch.ticketNumber
-      && !parsedBranch.description;
-  };
-
-  const isFeatureBranch = parsedBranch => {
-    return parsedBranch.version // TODO: semver-check dat
-      && parsedBranch.feature
-      && !parsedBranch.ticketNumber
-      && !parsedBranch.description;
-  };
-
-  const isDevBranch = parsedBranch => {
-    return parsedBranch.version // TODO: semver-check dat
-      && parsedBranch.description;
   };
 
   const getRepositoryOption = optionName => {
@@ -324,16 +143,16 @@ module.exports = (() => {
   };
 
   const getRemotes = () => {
-    const remotesAsString = execute(`git remote show`);
+    const remotesAsString = execution.execute(`git remote show`);
     return _(remotesAsString.split('\n'))
       .reject(remote => _.isEmpty(remote))
       .value();
   };
 
   const getBranchInfo = (branch, info) => {
-    const safeBranch = branch || getCurrentBranch();
+    const safeBranch = branch || branches.getCurrentBranch();
     try {
-      return execute(`git config branch.${safeBranch}.${info}`);
+      return execution.execute(`git config branch.${safeBranch}.${info}`);
     } catch (error) {
       return undefined;
     }
@@ -361,7 +180,6 @@ module.exports = (() => {
   };
 
   return {
-    BRANCH_TYPES,
     LOG_FORMATS,
     GIT_SERVERS_PRESET,
     GLOBAL_OPTIONS_FILE_PATH,
@@ -369,35 +187,17 @@ module.exports = (() => {
     REPOSITORY_OPTIONS_FILE_NAME,
 
     configureGutIfNeeded,
-
-    execute,
-    executeAndPipe,
-    print,
-    exit,
-
-    yesNoPrompt,
-    promisifiedPrompt,
+    getRepositoryOption,
 
     mergeArrayCustomizer,
 
-    getRepositoryOption,
-
-    getTopLevel,
     moveUpTop,
 
     isDirty,
     hasStagedChanges,
     hasUnstagedChanges,
 
-    getCurrentBranch,
-    searchForLocalBranch,
-    parseBranchName,
-    buildBranchName,
     getBranchOnlyCommits,
-    isMasterBranch,
-    isVersionBranch,
-    isFeatureBranch,
-    isDevBranch,
 
     getRemotes,
     getBranchInfo,
