+++ lib/utils/prompt.js
@@ -1,0 +1,32 @@
+module.exports = (() => {
+  const promptly = require('promptly');
+
+  const promisifiedPrompt = (message, options) => {
+    return new Promise((resolve, reject) => {
+      promptly.prompt(message, options, (error, value) => {
+        return error
+          ? reject(error)
+          : resolve(value);
+      });
+    });
+  };
+
+  const yesNoPrompt = (message, callback) => {
+    const options = {
+      default: 'n',
+      validator: choice => choice === 'y'
+    };
+    promptly.prompt(`${message} (y/n)`, options, (error, value) => {
+      if (error) {
+        throw error;
+      }
+
+      callback(value);
+    });
+  };
+
+  return {
+    yesNoPrompt,
+    promisifiedPrompt,
+  }
+})();
